requirements.txt
apache-airflow==2.6.2
psycopg2-binary==2.9.6
redis==4.5.5
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.4
